added file for customized testing companion.o config.h console.c console.h console.o csr.h device.c device.h device.o elf.c elf.h elf.o error.h excp.c excp.o ezheap.c ezheap.o fs.h get_init_fsimg.sh halt.c halt.h halt.o heap.h intr.c intr.h intr.o io.c io.h io.o kernel.elf kernel.ld kfs.c kfs.o kfs.raw main.c main.o main_shell.c main_tests.c main_test_vio.c Makefile memory.c memory.h memory.o mkcomp.sh plic.c plic.h plic.o process.c process.h process.o rubric_tests.c rubric_test.sh scnum.h start.o start.s stdlib.c stdlib.h string.c string.h string.o syscall.c syscall.o test_elf.c thrasm.o thrasm.s thread.c thread.h thread.o timer.c timer.h timer.o trapasm.o trapasm.s trap.h uart.c uart.h uart.o vioblk.c vioblk.o virtio.c virtio.h virtio.o, removed all warnings and possible errors

+++ src/kern/kfs.c
@@ -81,7 +81,7 @@ int fs_open(const char *name, struct io_intf **io)
       ioseek(fs_io, position);
       inode_t* file_inode = kmalloc(BLOCK_SIZE);
       ioread_full(fs_io, file_inode, BLOCK_SIZE);
-      uint64_t file_size =&file_inode->byte_len;
+      uint64_t file_size = (uint64_t)(&file_inode->byte_len);
       uint64_t flag = INUSE;
       for (int j = 0; j < MAX_FILE_OPEN; j++)
       {
